requirements.txt
streamlit==1.26.0
pandas==2.1.0
plotly==5.16.1
pdfplumber==0.10.1
python-docx==0.8.12
scikit-learn==1.3.2
torch==2.2.0
torchvision==0.17.1
sentence-transformers==2.2.2
